Added support for Artifactory 5.6
- The user "access-admin" is now ignored by this plugin, and will not be
  read or modified.
- Some checks to ensure proper password formats have been added.
- New getAllUsers API is now supported.
- New validate API has been added to ensure correct install/updates.
- New version and safety systems have been implemented.
- New fields are not yet supported.

+++ security/securityReplication/test/artifactory/SecurityTestApi.groovy
@@ -83,7 +83,7 @@ class SecurityTestApi {
     static def deleteAllUsers () {
         Collection<String> userNames = art.security().userNames()
         for (String userName : userNames) {
-            if (userName != "admin") {
+            if (userName != "anonymous" && userName != "admin" && userName != "access-admin") {
                 art.security().deleteUser(userName)
             }
         }
